Rename scripts and select for police
